Clarify immutability (#24)
It took me a few iterations of reading the spec to figure out that bytes
are indeed immutable, and not mutable, as implied by the last note of
HostLoadImportedModule. Similarly, in CreateBytesModule it's not
immediately obvious that the "immutable" descriptor applies to
SharedArrayBuffer as well.

As an unrelated bit of cleanup, the spec metadata is updated to match
the current stage of the proposal.